requirements.txt
click==8.1.3
colorlog==6.8.0
datasets==2.16.1
evidently==0.4.13
fastapi==0.109.0
hydra-core==1.3.2
Markdown==3.5.2
matplotlib==3.8.2
nltk==3.8.1
numpy==1.26.3
omegaconf==2.3.0
openai==0.28.0
pandas==2.1.4
protobuf==4.25.2
pydantic==1.10.13
pytest==7.4.4
rich==13.7.0
scikit_learn==1.3.2
tenacity==8.2.3
torch==2.1.2
transformers==4.36.2
accelerate>=0.20.1
scikit-learn==1.3.2
hydra-core==1.3.2
wandb==0.16.2
dvc==3.38.1
dvc-gs>=3.0.1
dvc-gdrive>=3.0.1
coverage
hydra_colorlog
python-multipart
click
markdown
evidently
nltk
matplotlib
hydra_colorlog==1.2.0
